add and delete numbers

+++ pantry/app/page.js
@@ -1,7 +1,7 @@
 'use client'
 import {Box, Stack, Typography, Button, Modal, TextField} from '@mui/material'
 import {firestore} from '@/firebase'
-import {collection, getDocs, query, doc, setDoc, deleteDoc} from 'firebase/firestore';
+import {collection, getDocs, query, doc, setDoc, deleteDoc, getDoc} from 'firebase/firestore';
 import {useEffect, useState} from 'react'
 
 
@@ -31,7 +31,7 @@ export default function Home() {
     const docs = await getDocs(snapshot)
     const pantryList = []
     docs.forEach((doc) => {
-      pantryList.push(doc.id)
+      pantryList.push({"name": doc.id, ... doc.data()})
     })
     console.log(pantryList)
     setPantry(pantryList)
@@ -43,14 +43,31 @@ export default function Home() {
 
   const addItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await setDoc(docRef, {})
-    updatePantry()
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+        const {count} = docSnap.data()
+        await setDoc(docRef, {count: count + 1})
+    }
+    else {
+      await setDoc(docRef, {count: 1})
+    }
+    await updatePantry()
   }
 
   const removeItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await deleteDoc(docRef)
-    updatePantry()
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      if (count == 1) {
+        await deleteDoc(docRef)
+      }
+      else
+      {
+        await setDoc(docRef, {count: count - 1})
+      }
+    }
+    await updatePantry()
   }
   return (
     <Box 
@@ -100,29 +117,33 @@ export default function Home() {
     </Box>
       <Stack width="800px" height="300px" spacing={2} overflow={'auto'}>
 
-        {pantry.map((i) => (
-        <Stack key = {i} direction={'row'} spacing={2} justifyContent={'center'} paddingX={2}alignItems={'space-between'}>
+        {pantry.map(({name, count}) => (
+        <Stack key = {name} direction={'row'} spacing={2} justifyContent={'center'} paddingX={2}alignItems={'space-between'}>
           <Box
-            key={i}
+            key={name}
             width="100%"
             minHeight="150px"
             display={'flex'}
             justifyContent={'center'}
             alignItems={'center'}
-            bgcolor={'#f0f0f0'}>
+            bgcolor={'#f0f0f0'}
+            gap={5}>
               <Typography
               variant={'h3'}
               color={'#333'}
               textAlign={'center'}>
                 {
                   //capitalize the first letter
-                  i.charAt(0).toUpperCase() + i.slice(1)
+                  name.charAt(0).toUpperCase() + name.slice(1)
                 }
+              </Typography>
+              <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+                  Quantity: {count}
               </Typography>
               
             </Box>
             <Button variant="contained" onClick={() => 
-              removeItem(i)
+              removeItem(name)
             }>Delete</Button>
    
           </Stack>
